migrate to tailwind and update folder structure
