--- src/components/MovieCard.tsx
@@ -1,41 +1,0 @@
-import "../css/MovieCard.css";
-import { useMovieContext } from "../contexts/MovieContext";
-import { Movie } from "../types/movie";
-import React from "react";
-
-export default function MovieCard({ movie }: { movie: Movie }) {
-  const { isFavourite, addToFavourites, removeFromFavourites } =
-    useMovieContext();
-  const favourite = isFavourite(movie.id);
-
-  function onFavouriteClick(e: React.MouseEvent<HTMLButtonElement>) {
-    e.preventDefault();
-    if (favourite) removeFromFavourites(movie.id);
-    else addToFavourites(movie);
-  }
-
-  return (
-    <>
-      <div className="movie-card">
-        <div className="movie-poster">
-          <img
-            src={`https://image.tmdb.org/t/p/w500${movie.poster_path}`}
-            alt={movie.title}
-          />
-          <div className="movie-overlay">
-            <button
-              className={favourite ? "favourite-btn active" : "favourite-btn"}
-              onClick={onFavouriteClick}
-            >
-              &hearts;
-            </button>
-          </div>
-        </div>
-        <div className="movie-info">
-          <h3>{movie.title}</h3>
-          <p>{movie.release_date?.split("-")[0]}</p>
-        </div>
-      </div>
-    </>
-  );
-}
