--- src/css/App.css
@@ -1,37 +1,0 @@
-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 1000;
-  background: linear-gradient(
-    to bottom,
-    rgba(0, 0, 0, 0.7) 10%,
-    rgba(0, 0, 0, 0)
-  );
-  transition: background-color 0.3s ease;
-}
-
-header.scrolled {
-  background-color: var(--netflix-black);
-}
-
-.main-content {
-  min-height: 100vh;
-  padding-top: 80px;
-  padding-bottom: 40px;
-}
-
-@media (max-width: 768px) {
-  .main-content {
-    padding-top: 70px;
-    padding-bottom: 30px;
-  }
-}
-
-@media (max-width: 480px) {
-  .main-content {
-    padding-top: 60px;
-    padding-bottom: 20px;
-  }
-}
