--- src/css/Favourites.css
@@ -1,90 +1,0 @@
-.favourites {
-  max-width: 1400px;
-  margin: 0 auto;
-  padding: 0 50px;
-}
-
-.favourites-empty {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 60vh;
-  text-align: center;
-  padding: 40px 20px;
-}
-
-.favourites-empty h2 {
-  font-size: 2rem;
-  color: var(--netflix-white);
-  margin-bottom: 15px;
-  font-weight: 600;
-}
-
-.favourites-empty p {
-  font-size: 1.1rem;
-  color: var(--netflix-light-gray);
-  max-width: 500px;
-  line-height: 1.6;
-}
-
-.movies-grid {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  gap: 20px;
-  margin-top: 40px;
-}
-
-@media (max-width: 1400px) {
-  .favourites .movies-grid {
-    grid-template-columns: repeat(4, 1fr);
-  }
-}
-
-@media (max-width: 1024px) {
-  .favourites .movies-grid {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-
-@media (max-width: 768px) {
-  .favourites {
-    padding: 0 15px;
-  }
-
-  .favourites .movies-grid {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 12px;
-  }
-
-  .favourites-empty {
-    padding: 30px 15px;
-  }
-
-  .favourites-empty h2 {
-    font-size: 1.5rem;
-  }
-
-  .favourites-empty p {
-    font-size: 1rem;
-    padding: 0 10px;
-  }
-}
-
-@media (max-width: 480px) {
-  .favourites {
-    padding: 0 10px;
-  }
-
-  .favourites .movies-grid {
-    gap: 10px;
-  }
-
-  .favourites-empty h2 {
-    font-size: 1.3rem;
-  }
-
-  .favourites-empty p {
-    font-size: 0.9rem;
-  }
-}
