--- src/css/Home.css
@@ -1,145 +1,0 @@
-.home {
-  max-width: 1400px;
-  margin: 0 auto;
-  padding: 0 50px;
-}
-
-.search-form {
-  display: flex;
-  gap: 15px;
-  margin-bottom: 40px;
-  max-width: 800px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.search-form input {
-  flex: 1;
-  padding: 15px 20px;
-  background-color: rgba(255, 255, 255, 0.1);
-  border: 2px solid rgba(255, 255, 255, 0.2);
-  border-radius: 4px;
-  color: var(--netflix-white);
-  font-size: 1rem;
-  outline: none;
-  transition: all 0.3s ease;
-}
-
-.search-form input::placeholder {
-  color: var(--netflix-gray);
-}
-
-.search-form input:focus {
-  background-color: rgba(255, 255, 255, 0.15);
-  border-color: var(--netflix-white);
-}
-
-.search-btn {
-  padding: 15px 35px;
-  background-color: var(--netflix-red);
-  color: var(--netflix-white);
-  font-size: 1rem;
-  font-weight: 600;
-  border-radius: 4px;
-  transition: all 0.3s ease;
-}
-
-.search-btn:hover {
-  background-color: var(--netflix-dark-red);
-  transform: scale(1.05);
-}
-
-.search-btn:active {
-  transform: scale(0.98);
-}
-
-.movies-grid {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  gap: 20px;
-  margin-top: 40px;
-}
-
-.error-message {
-  color: var(--netflix-red);
-  text-align: center;
-  font-size: 1.2rem;
-  margin: 20px 0;
-}
-
-.loading {
-  color: var(--netflix-white);
-  text-align: center;
-  font-size: 1.5rem;
-  margin: 40px 0;
-}
-
-.favourites-title {
-  display: flex;
-  justify-content: center;
-}
-
-@media (max-width: 1400px) {
-  .movies-grid {
-    grid-template-columns: repeat(4, 1fr);
-  }
-}
-
-@media (max-width: 1024px) {
-  .movies-grid {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-
-@media (max-width: 768px) {
-  .home {
-    padding: 0 15px;
-  }
-
-  .movies-grid {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 12px;
-  }
-
-  .search-form {
-    flex-direction: column;
-    gap: 12px;
-    margin-bottom: 25px;
-  }
-
-  .search-form input {
-    padding: 14px 18px;
-    font-size: 16px;
-  }
-
-  .search-btn {
-    padding: 14px 24px;
-    font-size: 16px;
-    width: 100%;
-  }
-
-  .error-message {
-    font-size: 1rem;
-    padding: 0 10px;
-  }
-
-  .loading {
-    font-size: 1.2rem;
-    margin: 30px 0;
-  }
-}
-
-@media (max-width: 480px) {
-  .home {
-    padding: 0 10px;
-  }
-
-  .movies-grid {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 10px;
-  }
-
-  .search-form {
-    margin-bottom: 20px;
-  }
-}
