--- src/css/MovieCard.css
@@ -1,163 +1,0 @@
-.movie-card {
-  border-radius: 8px;
-  overflow: hidden;
-  transition: all 0.3s ease;
-  cursor: pointer;
-  background-color: rgba(255, 255, 255, 0.05);
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.movie-card:hover {
-  transform: scale(1.08);
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.6);
-  z-index: 10;
-}
-
-.movie-poster {
-  position: relative;
-  overflow: hidden;
-  width: 100%;
-  aspect-ratio: 2 / 3;
-}
-
-.movie-poster img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  display: block;
-  transition: all 0.3s ease;
-}
-
-.movie-card:hover .movie-poster img {
-  filter: brightness(0.7);
-}
-
-.movie-overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: linear-gradient(
-    to bottom,
-    transparent 0%,
-    rgba(0, 0, 0, 0.8) 100%
-  );
-  opacity: 0;
-  transition: opacity 0.3s ease;
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  padding: 20px;
-}
-
-.movie-card:hover .movie-overlay {
-  opacity: 1;
-}
-
-.favourite-btn {
-  background-color: transparent;
-  border: 2px solid var(--netflix-white);
-  color: var(--netflix-white);
-  font-size: 1.5rem;
-  padding: 8px 16px;
-  border-radius: 50%;
-  width: 50px;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.3s ease;
-}
-
-.favourite-btn:hover {
-  background-color: var(--netflix-red);
-  border-color: var(--netflix-red);
-  transform: scale(1.1);
-}
-
-.favourite-btn.active {
-  background-color: var(--netflix-red);
-  border-color: var(--netflix-red);
-}
-
-.movie-info {
-  padding: 15px;
-  background-color: rgba(0, 0, 0, 0.6);
-  flex-shrink: 0;
-}
-
-.movie-info h3 {
-  font-size: 1rem;
-  font-weight: 600;
-  margin-bottom: 5px;
-  color: var(--netflix-white);
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.movie-info p {
-  font-size: 0.875rem;
-  color: var(--netflix-light-gray);
-  margin: 0;
-}
-
-@media (max-width: 768px) {
-  .movie-card:hover {
-    transform: scale(1.05);
-  }
-
-  .movie-info {
-    padding: 10px;
-  }
-
-  .movie-info h3 {
-    font-size: 0.85rem;
-  }
-
-  .movie-info p {
-    font-size: 0.75rem;
-  }
-
-  .favourite-btn {
-    width: 45px;
-    height: 45px;
-    font-size: 1.3rem;
-  }
-
-  .movie-overlay {
-    padding: 15px;
-  }
-}
-
-@media (max-width: 480px) {
-  .movie-card:hover {
-    transform: scale(1.03);
-  }
-
-  .movie-info {
-    padding: 8px;
-  }
-
-  .movie-info h3 {
-    font-size: 0.8rem;
-  }
-
-  .movie-info p {
-    font-size: 0.7rem;
-  }
-
-  .favourite-btn {
-    width: 40px;
-    height: 40px;
-    font-size: 1.2rem;
-  }
-
-  .movie-overlay {
-    padding: 10px;
-    opacity: 1;
-  }
-}
