--- src/css/NavBar.css
@@ -1,88 +1,0 @@
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px 50px;
-  width: 100%;
-}
-
-.navbar-brand {
-  font-size: 1.8rem;
-  font-weight: bold;
-  color: var(--netflix-red);
-  letter-spacing: 2px;
-}
-
-.navbar-brand a {
-  color: var(--netflix-red);
-}
-
-.navbar-brand a:hover {
-  color: var(--netflix-dark-red);
-}
-
-.navbar-links {
-  display: flex;
-  gap: 30px;
-  align-items: center;
-}
-
-.nav-link {
-  color: var(--netflix-light-gray);
-  font-size: 1rem;
-  font-weight: 500;
-  padding: 8px 16px;
-  border-radius: 4px;
-  transition: color 0.3s ease, background-color 0.3s ease;
-}
-
-.nav-link:hover {
-  color: var(--netflix-white);
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.nav-link.active {
-  color: var(--netflix-white);
-  font-weight: 600;
-}
-
-@media (max-width: 768px) {
-  .navbar {
-    padding: 12px 15px;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-
-  .navbar-brand {
-    font-size: 1.4rem;
-    letter-spacing: 1px;
-  }
-
-  .navbar-links {
-    gap: 12px;
-  }
-
-  .nav-link {
-    font-size: 0.9rem;
-    padding: 8px 12px;
-  }
-}
-
-@media (max-width: 480px) {
-  .navbar {
-    padding: 10px;
-  }
-
-  .navbar-brand {
-    font-size: 1.2rem;
-  }
-
-  .navbar-links {
-    gap: 8px;
-  }
-
-  .nav-link {
-    font-size: 0.85rem;
-    padding: 6px 10px;
-  }
-}
