+++ src/css/index.css
@@ -1,62 +1,22 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
+@import 'tailwindcss';
 
-:root {
-  --netflix-black: #141414;
-  --netflix-red: #e50914;
-  --netflix-dark-red: #b20710;
-  --netflix-white: #ffffff;
-  --netflix-gray: #808080;
-  --netflix-light-gray: #b3b3b3;
+@theme {
+  --color-netflix-black: #141414;
+  --color-netflix-red: #e50914;
+  --color-netflix-dark-red: #b20710;
+  --color-netflix-white: #ffffff;
+  --color-netflix-gray: #808080;
+  --color-netflix-light-gray: #b3b3b3;
+  --color-netflix-dark-gray: #2f2f2f;
+  --color-netflix-hover: #181818;
+
+  --font-family-sans: 'Helvetica Neue', Helvetica, Arial, sans-serif;
 }
 
 body {
-  margin: 0;
-  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-  background-color: var(--netflix-black);
-  color: var(--netflix-white);
-  min-height: 100vh;
-  line-height: 1.5;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+  @apply bg-netflix-black text-netflix-white min-h-screen font-sans;
 }
 
 #root {
-  min-height: 100vh;
-  width: 100%;
-}
-
-a {
-  text-decoration: none;
-  color: inherit;
-  transition: color 0.3s ease;
-}
-
-button {
-  font-family: inherit;
-  cursor: pointer;
-  border: none;
-  outline: none;
-  transition: all 0.3s ease;
-}
-
-img {
-  max-width: 100%;
-  display: block;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-weight: 600;
-}
-
-input {
-  font-family: inherit;
+  @apply min-h-screen bg-netflix-black;
 }
