+++ src/features/favourites/index.ts
@@ -1,0 +1,1 @@
+export { default as Favourites } from './Favourites';
