+++ src/features/movies/index.ts
@@ -1,0 +1,4 @@
+export { default as MovieCard } from './MovieCard';
+export { default as Home } from './Home';
+export * from './movie.types';
+export * from './movies.api';
