--- src/pages/Home.tsx
@@ -1,43 +1,0 @@
-import MovieCard from "../components/MovieCard";
-import { useState, useEffect } from "react";
-import { getPopularMovies } from "../services/api";
-import { Movie } from "../types/movie";
-import "../css/Home.css";
-
-export default function Home() {
-  const [movies, setMovies] = useState<Movie[]>([]);
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-
-  useEffect(() => {
-    const loadPopularMovies = async () => {
-      try {
-        const popularMovies = await getPopularMovies();
-        setMovies(popularMovies);
-      } catch (err) {
-        console.log(err);
-        setError("Failed to load movies...");
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    loadPopularMovies();
-  }, []);
-
-  return (
-    <div className="home">
-      <h1>Trending Movies</h1>
-      {error && <div className="error-message">{error} </div>}
-      {loading ? (
-        <div className="loading">Loading...</div>
-      ) : (
-        <div className="movies-grid">
-          {movies.map((movie) => (
-            <MovieCard movie={movie} key={movie.id} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
