--- src/pages/Search.tsx
@@ -1,77 +1,0 @@
-import MovieCard from "../components/MovieCard";
-import { useState, useEffect } from "react";
-import { searchMovies, getPopularMovies } from "../services/api";
-import { Movie } from "../types/movie";
-import React from "react";
-import "../css/Home.css";
-
-export default function Search() {
-  const [searchQuery, setSearchQuery] = useState<string>("");
-  const [movies, setMovies] = useState<Movie[]>([]);
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-
-  const handleSearch = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!searchQuery.trim()) return;
-    if (loading) return;
-
-    setLoading(true);
-
-    try {
-      const searchResults = await searchMovies(searchQuery);
-      setMovies(searchResults);
-      setError(null);
-    } catch (err) {
-      setError("Failed to search movies...")
-      console.log(err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-    useEffect(() => {
-    const loadPopularMovies = async () => {
-      try {
-        const popularMovies = await getPopularMovies();
-        setMovies(popularMovies);
-      } catch (err) {
-        console.log(err);
-        setError("Failed to load movies...");
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    loadPopularMovies();
-  }, []);
-
-  return (
-    <div className="home">
-      <form onSubmit={handleSearch} className="search-form">
-        <label htmlFor="input-form"></label>
-        <input
-          type="text"
-          name="search-input"
-          id="input"
-          placeholder="Search for movies..."
-          value={searchQuery}
-          onChange={(e) => setSearchQuery(e.target.value)}
-        />
-        <button type="submit" className="search-btn">
-          Search
-        </button>
-      </form>
-      {error && <div className="error-message">{error} </div>}
-      {loading ? (
-        <div className="loading">Loading...</div>
-      ) : (
-        <div className="movies-grid">
-          {movies.map((movie) => (
-            <MovieCard movie={movie} key={movie.id} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
